requirements.txt
# RAG framework
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-ollama

# Vector database
chromadb

# Embeddings
sentence-transformers

# Data collection
nflreadpy

# Utilities
python-dotenv

# Testing
pytest
